requirements.txt
pandas
matplotlib
seaborn
plotly
statsmodels
pywaffle
numpy
hiplot
pillow
networkx
streamlit_option_menu
scikit-learn
tensorflow
